# Fast softmax